fix: address PR review — unused imports, shouldRepaint staleness, clamp int type, CalendarExportHelper tests (#83)
* Initial plan

* fix: address PR review feedback - unused imports, shouldRepaint, clamp type, add CalendarExportHelper tests


---------

+++ frontend/lib/features/agenda/presentation/pages/agenda_page.dart
@@ -176,7 +176,7 @@ class _AgendaPageContentState extends State<_AgendaPageContent> {
                   Row(
                     mainAxisSize: MainAxisSize.min,
                     children: List.generate(
-                      dayEvents.length.clamp(0, 3),
+                      dayEvents.length.clamp(0, 3).toInt(),
                       (_) => Container(
                         width: 5,
                         height: 5,
